Move old JLDSerializer out of the way
