+++ src/main/scala/org/clulab/wm/eidos/apps/ExtractFromDirectory.scala
@@ -4,7 +4,7 @@ import java.io.PrintWriter
 import scala.collection.Seq
 import org.clulab.serialization.json.stringify
 import org.clulab.wm.eidos.EidosSystem
-import org.clulab.wm.eidos.serialization.json.JLDCorpus
+import org.clulab.wm.eidos.serialization.json.odin.JLDCorpus
 import org.clulab.wm.eidos.utils.FileUtils.findFiles
 
 object ExtractFromDirectory extends App {
